feat(parser): refine graph extraction and cleanup algorithms
- Remove Floyd–Warshall from algorithm detection and keyword rules

- Split Euler/Hamilton into problem_group: euler_hamilton

- Filter noisy vertices/edges and remove self-loops

- Improve DFS detection for " duyet do thi theo chieu sau\

+++ Models/Edge.cs
@@ -4,5 +4,5 @@ public sealed class Edge
 {
     public string from { get; set; } = "";
     public string to { get; set; } = "";
-    public int? weight { get; set; }
+    public double? weight { get; set; }
 }
